backend/requirements.txt
fastapi
fastapi[standard]
fastapi-nextauth-jwt # to verify tokens from frontend
uvicorn
python-dotenv
